move dcmread to utils
