+++ deid/dicom/header.py
@@ -5,10 +5,10 @@ __license__ = "MIT"
 
 import os
 
-from deid.dicom.dcmread import dcmread
 from deid.dicom.parser import DicomParser
 from deid.dicom.utils import save_dicom
 from deid.logger import bot
+from deid.utils.dcmread import dcmread
 
 here = os.path.dirname(os.path.abspath(__file__))
