+++ deid/dicom/parser.py
@@ -14,13 +14,13 @@ from pydicom.tag import Tag
 from deid.config import DeidRecipe
 from deid.config.standards import actions as valid_actions
 from deid.dicom.actions import deid_funcs, jitter_timestamp
-from deid.dicom.dcmread import dcmread
 from deid.dicom.fields import DicomField, expand_field_expression, get_fields
 from deid.dicom.groups import extract_fields_list, extract_values_list
 from deid.dicom.tags import add_tag, get_private, get_tag, remove_sequences
 from deid.dicom.utils import save_dicom
 from deid.logger import bot
 from deid.utils import parse_value, read_json
+from deid.utils.dcmread import dcmread
 
 here = os.path.dirname(os.path.abspath(__file__))
