+++ deid/dicom/pixels/clean.py
@@ -17,9 +17,9 @@ from pydicom.pixel_data_handlers.util import get_expected_length
 
 from deid.config import DeidRecipe
 from deid.dicom import utils
-from deid.dicom.dcmread import dcmread
 from deid.logger import bot
 from deid.utils import get_temporary_name
+from deid.utils.dcmread import dcmread
 
 matplotlib.use("pdf")
