+++ deid/dicom/utils.py
@@ -10,8 +10,8 @@ from pydicom import FileDataset
 
 from deid.logger import bot
 from deid.utils import recursive_find
+from deid.utils.dcmread import dcmread
 
-from .dcmread import dcmread
 from .validate import validate_dicoms
 
 ################################################################################
