+++ deid/tests/test_clean.py
@@ -13,10 +13,10 @@ from copy import deepcopy
 
 from deid.config import DeidRecipe
 from deid.data import get_dataset
-from deid.dicom.dcmread import dcmread
 from deid.dicom.pixels import clean_pixel_data, has_burned_pixels
 from deid.tests.common import get_file
 from deid.utils import get_installdir
+from deid.utils.dcmread import dcmread
 
 global generate_uid
