+++ deid/tests/test_sequence_remove.py
@@ -11,9 +11,9 @@ import unittest
 
 from deid.data import get_dataset
 from deid.dicom import get_files, replace_identifiers
-from deid.dicom.dcmread import dcmread
 from deid.tests.common import create_recipe
 from deid.utils import get_installdir
+from deid.utils.dcmread import dcmread
 
 global generate_uid
